refactor(ext/ffi): split into multiple components
