+++ ext/ffi/lib.rs
@@ -2,8 +2,6 @@
 
 use core::ptr::NonNull;
 use deno_core::anyhow::anyhow;
-use deno_core::error::generic_error;
-use deno_core::error::range_error;
 use deno_core::error::type_error;
 use deno_core::error::AnyError;
 use deno_core::futures::channel::mpsc;
@@ -19,21 +17,18 @@ use deno_core::Extension;
 use deno_core::OpState;
 use deno_core::Resource;
 use deno_core::ResourceId;
-use dlopen::raw::Library;
+
 use libffi::middle::Arg;
 use libffi::middle::Cif;
 use serde::Deserialize;
 use std::borrow::Cow;
 use std::cell::RefCell;
-use std::collections::HashMap;
 use std::ffi::c_void;
-use std::ffi::CStr;
 use std::future::IntoFuture;
 use std::mem::size_of;
 use std::os::raw::c_char;
 use std::os::raw::c_short;
 use std::path::Path;
-use std::path::PathBuf;
 use std::pin::Pin;
 use std::ptr;
 use std::rc::Rc;
@@ -41,7 +36,14 @@ use std::sync::mpsc::sync_channel;
 use std::task::Poll;
 use std::task::Waker;
 
+mod dlfcn;
 mod fast_call;
+mod repr;
+mod symbol;
+
+use dlfcn::{op_ffi_load, DynamicLibraryResource, ForeignFunction};
+use repr::*;
+use symbol::{NativeType, Symbol};
 
 #[cfg(not(target_pointer_width = "64"))]
 compile_error!("platform not supported");
@@ -83,21 +85,6 @@ pub trait FfiPermissions {
 }
 
 #[derive(Clone)]
-struct Symbol {
-  cif: libffi::middle::Cif,
-  ptr: libffi::middle::CodePtr,
-  parameter_types: Vec<NativeType>,
-  result_type: NativeType,
-  can_callback: bool,
-}
-
-#[allow(clippy::non_send_fields_in_send_ty)]
-// SAFETY: unsafe trait must have unsafe implementation
-unsafe impl Send for Symbol {}
-// SAFETY: unsafe trait must have unsafe implementation
-unsafe impl Sync for Symbol {}
-
-#[derive(Clone)]
 struct PtrSymbol {
   cif: libffi::middle::Cif,
   ptr: libffi::middle::CodePtr,
@@ -124,38 +111,6 @@ impl PtrSymbol {
 unsafe impl Send for PtrSymbol {}
 // SAFETY: unsafe trait must have unsafe implementation
 unsafe impl Sync for PtrSymbol {}
-
-struct DynamicLibraryResource {
-  lib: Library,
-  symbols: HashMap<String, Box<Symbol>>,
-}
-
-impl Resource for DynamicLibraryResource {
-  fn name(&self) -> Cow<str> {
-    "dynamicLibrary".into()
-  }
-
-  fn close(self: Rc<Self>) {
-    drop(self)
-  }
-}
-
-impl DynamicLibraryResource {
-  fn get_static(&self, symbol: String) -> Result<*const c_void, AnyError> {
-    // By default, Err returned by this function does not tell
-    // which symbol wasn't exported. So we'll modify the error
-    // message to include the name of symbol.
-    //
-    // SAFETY: The obtained T symbol is the size of a pointer.
-    match unsafe { self.lib.symbol::<*const c_void>(&symbol) } {
-      Ok(value) => Ok(Ok(value)),
-      Err(err) => Err(generic_error(format!(
-        "Failed to register symbol {}: {}",
-        symbol, err
-      ))),
-    }?
-  }
-}
 
 type PendingFfiAsyncWork = Box<dyn FnOnce()>;
 
@@ -236,53 +191,6 @@ pub fn init<P: FfiPermissions + 'static>(unstable: bool) -> Extension {
       Ok(())
     })
     .build()
-}
-
-/// Defines the accepted types that can be used as
-/// parameters and return values in FFI.
-#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "lowercase")]
-enum NativeType {
-  Void,
-  Bool,
-  U8,
-  I8,
-  U16,
-  I16,
-  U32,
-  I32,
-  U64,
-  I64,
-  USize,
-  ISize,
-  F32,
-  F64,
-  Pointer,
-  Buffer,
-  Function,
-}
-
-impl From<NativeType> for libffi::middle::Type {
-  fn from(native_type: NativeType) -> Self {
-    match native_type {
-      NativeType::Void => libffi::middle::Type::void(),
-      NativeType::U8 | NativeType::Bool => libffi::middle::Type::u8(),
-      NativeType::I8 => libffi::middle::Type::i8(),
-      NativeType::U16 => libffi::middle::Type::u16(),
-      NativeType::I16 => libffi::middle::Type::i16(),
-      NativeType::U32 => libffi::middle::Type::u32(),
-      NativeType::I32 => libffi::middle::Type::i32(),
-      NativeType::U64 => libffi::middle::Type::u64(),
-      NativeType::I64 => libffi::middle::Type::i64(),
-      NativeType::USize => libffi::middle::Type::usize(),
-      NativeType::ISize => libffi::middle::Type::isize(),
-      NativeType::F32 => libffi::middle::Type::f32(),
-      NativeType::F64 => libffi::middle::Type::f64(),
-      NativeType::Pointer | NativeType::Buffer | NativeType::Function => {
-        libffi::middle::Type::pointer()
-      }
-    }
-  }
 }
 
 /// Intermediate format for easy translation from NativeType + V8 value
@@ -466,333 +374,6 @@ impl NativeValue {
 
 // SAFETY: unsafe trait must have unsafe implementation
 unsafe impl Send for NativeValue {}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-struct ForeignFunction {
-  name: Option<String>,
-  parameters: Vec<NativeType>,
-  result: NativeType,
-  #[serde(rename = "nonblocking")]
-  non_blocking: Option<bool>,
-  #[serde(rename = "callback")]
-  #[serde(default = "default_callback")]
-  callback: bool,
-}
-
-fn default_callback() -> bool {
-  false
-}
-
-// ForeignStatic's name and type fields are read and used by
-// serde_v8 to determine which variant a ForeignSymbol is.
-// They are not used beyond that and are thus marked with underscores.
-#[derive(Deserialize, Debug)]
-struct ForeignStatic {
-  #[serde(rename(deserialize = "name"))]
-  _name: Option<String>,
-  #[serde(rename(deserialize = "type"))]
-  _type: String,
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(untagged)]
-enum ForeignSymbol {
-  ForeignFunction(ForeignFunction),
-  ForeignStatic(ForeignStatic),
-}
-
-#[derive(Deserialize, Debug)]
-struct FfiLoadArgs {
-  path: String,
-  symbols: HashMap<String, ForeignSymbol>,
-}
-
-// `path` is only used on Windows.
-#[allow(unused_variables)]
-pub(crate) fn format_error(e: dlopen::Error, path: String) -> String {
-  match e {
-    #[cfg(target_os = "windows")]
-    // This calls FormatMessageW with library path
-    // as replacement for the insert sequences.
-    // Unlike libstd which passes the FORMAT_MESSAGE_IGNORE_INSERTS
-    // flag without any arguments.
-    //
-    // https://github.com/denoland/deno/issues/11632
-    dlopen::Error::OpeningLibraryError(e) => {
-      use std::ffi::OsStr;
-      use std::os::windows::ffi::OsStrExt;
-      use winapi::shared::minwindef::DWORD;
-      use winapi::shared::winerror::ERROR_INSUFFICIENT_BUFFER;
-      use winapi::um::errhandlingapi::GetLastError;
-      use winapi::um::winbase::FormatMessageW;
-      use winapi::um::winbase::FORMAT_MESSAGE_ARGUMENT_ARRAY;
-      use winapi::um::winbase::FORMAT_MESSAGE_FROM_SYSTEM;
-      use winapi::um::winnt::LANG_SYSTEM_DEFAULT;
-      use winapi::um::winnt::MAKELANGID;
-      use winapi::um::winnt::SUBLANG_SYS_DEFAULT;
-
-      let err_num = match e.raw_os_error() {
-        Some(err_num) => err_num,
-        // This should never hit unless dlopen changes its error type.
-        None => return e.to_string(),
-      };
-
-      // Language ID (0x0800)
-      let lang_id =
-        MAKELANGID(LANG_SYSTEM_DEFAULT, SUBLANG_SYS_DEFAULT) as DWORD;
-
-      let mut buf = vec![0; 500];
-
-      let path = OsStr::new(&path)
-        .encode_wide()
-        .chain(Some(0).into_iter())
-        .collect::<Vec<_>>();
-
-      let arguments = [path.as_ptr()];
-
-      loop {
-        // SAFETY:
-        // winapi call to format the error message
-        let length = unsafe {
-          FormatMessageW(
-            FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_ARGUMENT_ARRAY,
-            std::ptr::null_mut(),
-            err_num as DWORD,
-            lang_id as DWORD,
-            buf.as_mut_ptr(),
-            buf.len() as DWORD,
-            arguments.as_ptr() as _,
-          )
-        };
-
-        if length == 0 {
-          // SAFETY:
-          // winapi call to get the last error message
-          let err_num = unsafe { GetLastError() };
-          if err_num == ERROR_INSUFFICIENT_BUFFER {
-            buf.resize(buf.len() * 2, 0);
-            continue;
-          }
-
-          // Something went wrong, just return the original error.
-          return e.to_string();
-        }
-
-        let msg = String::from_utf16_lossy(&buf[..length as usize]);
-        return msg;
-      }
-    }
-    _ => e.to_string(),
-  }
-}
-
-#[op(v8)]
-fn op_ffi_load<FP, 'scope>(
-  scope: &mut v8::HandleScope<'scope>,
-  state: &mut deno_core::OpState,
-  args: FfiLoadArgs,
-) -> Result<(ResourceId, serde_v8::Value<'scope>), AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  let path = args.path;
-
-  check_unstable(state, "Deno.dlopen");
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(Some(&PathBuf::from(&path)))?;
-
-  let lib = Library::open(&path).map_err(|e| {
-    dlopen::Error::OpeningLibraryError(std::io::Error::new(
-      std::io::ErrorKind::Other,
-      format_error(e, path),
-    ))
-  })?;
-  let mut resource = DynamicLibraryResource {
-    lib,
-    symbols: HashMap::new(),
-  };
-  let obj = v8::Object::new(scope);
-
-  for (symbol_key, foreign_symbol) in args.symbols {
-    match foreign_symbol {
-      ForeignSymbol::ForeignStatic(_) => {
-        // No-op: Statics will be handled separately and are not part of the Rust-side resource.
-      }
-      ForeignSymbol::ForeignFunction(foreign_fn) => {
-        let symbol = match &foreign_fn.name {
-          Some(symbol) => symbol,
-          None => &symbol_key,
-        };
-        // By default, Err returned by this function does not tell
-        // which symbol wasn't exported. So we'll modify the error
-        // message to include the name of symbol.
-        let fn_ptr =
-          // SAFETY: The obtained T symbol is the size of a pointer.
-          match unsafe { resource.lib.symbol::<*const c_void>(symbol) } {
-            Ok(value) => Ok(value),
-            Err(err) => Err(generic_error(format!(
-              "Failed to register symbol {}: {}",
-              symbol, err
-            ))),
-          }?;
-        let ptr = libffi::middle::CodePtr::from_ptr(fn_ptr as _);
-        let cif = libffi::middle::Cif::new(
-          foreign_fn
-            .parameters
-            .clone()
-            .into_iter()
-            .map(libffi::middle::Type::from),
-          foreign_fn.result.into(),
-        );
-
-        let func_key = v8::String::new(scope, &symbol_key).unwrap();
-        let sym = Box::new(Symbol {
-          cif,
-          ptr,
-          parameter_types: foreign_fn.parameters,
-          result_type: foreign_fn.result,
-          can_callback: foreign_fn.callback,
-        });
-
-        resource.symbols.insert(symbol_key, sym.clone());
-        match foreign_fn.non_blocking {
-          // Generate functions for synchronous calls.
-          Some(false) | None => {
-            let function = make_sync_fn(scope, sym);
-            obj.set(scope, func_key.into(), function.into());
-          }
-          // This optimization is not yet supported for non-blocking calls.
-          _ => {}
-        };
-      }
-    }
-  }
-
-  let rid = state.resource_table.add(resource);
-  Ok((
-    rid,
-    serde_v8::Value {
-      v8_value: obj.into(),
-    },
-  ))
-}
-
-fn needs_unwrap(rv: NativeType) -> bool {
-  matches!(
-    rv,
-    NativeType::Function
-      | NativeType::Pointer
-      | NativeType::Buffer
-      | NativeType::I64
-      | NativeType::ISize
-      | NativeType::U64
-      | NativeType::USize
-  )
-}
-
-fn is_i64(rv: NativeType) -> bool {
-  matches!(rv, NativeType::I64 | NativeType::ISize)
-}
-
-// Create a JavaScript function for synchronous FFI call to
-// the given symbol.
-fn make_sync_fn<'s>(
-  scope: &mut v8::HandleScope<'s>,
-  sym: Box<Symbol>,
-) -> v8::Local<'s, v8::Function> {
-  let sym = Box::leak(sym);
-  let builder = v8::FunctionTemplate::builder(
-    |scope: &mut v8::HandleScope,
-     args: v8::FunctionCallbackArguments,
-     mut rv: v8::ReturnValue| {
-      let external: v8::Local<v8::External> = args.data().try_into().unwrap();
-      // SAFETY: The pointer will not be deallocated until the function is
-      // garbage collected.
-      let symbol = unsafe { &*(external.value() as *const Symbol) };
-      let needs_unwrap = match needs_unwrap(symbol.result_type) {
-        true => Some(args.get(symbol.parameter_types.len() as i32)),
-        false => None,
-      };
-      match ffi_call_sync(scope, args, symbol) {
-        Ok(result) => {
-          match needs_unwrap {
-            Some(v) => {
-              let view: v8::Local<v8::ArrayBufferView> = v.try_into().unwrap();
-              let backing_store =
-                view.buffer(scope).unwrap().get_backing_store();
-
-              if is_i64(symbol.result_type) {
-                // SAFETY: v8::SharedRef<v8::BackingStore> is similar to Arc<[u8]>,
-                // it points to a fixed continuous slice of bytes on the heap.
-                let bs = unsafe {
-                  &mut *(&backing_store[..] as *const _ as *mut [u8]
-                    as *mut i64)
-                };
-                // SAFETY: We already checked that type == I64
-                let value = unsafe { result.i64_value };
-                *bs = value;
-              } else {
-                // SAFETY: v8::SharedRef<v8::BackingStore> is similar to Arc<[u8]>,
-                // it points to a fixed continuous slice of bytes on the heap.
-                let bs = unsafe {
-                  &mut *(&backing_store[..] as *const _ as *mut [u8]
-                    as *mut u64)
-                };
-                // SAFETY: We checked that type == U64
-                let value = unsafe { result.u64_value };
-                *bs = value;
-              }
-            }
-            None => {
-              // SAFETY: Same return type declared to libffi; trust user to have it right beyond that.
-              let result = unsafe { result.to_v8(scope, symbol.result_type) };
-              rv.set(result.v8_value);
-            }
-          }
-        }
-        Err(err) => {
-          deno_core::_ops::throw_type_error(scope, err.to_string());
-        }
-      };
-    },
-  )
-  .data(v8::External::new(scope, sym as *mut Symbol as *mut _).into());
-
-  let mut fast_call_alloc = None;
-
-  let func = if fast_call::is_compatible(sym) {
-    let trampoline = fast_call::compile_trampoline(sym);
-    let func = builder.build_fast(
-      scope,
-      &fast_call::make_template(sym, &trampoline),
-      None,
-    );
-    fast_call_alloc = Some(Box::into_raw(Box::new(trampoline)));
-    func
-  } else {
-    builder.build(scope)
-  };
-  let func = func.get_function(scope).unwrap();
-
-  let weak = v8::Weak::with_finalizer(
-    scope,
-    func,
-    Box::new(move |_| {
-      // SAFETY: This is never called twice. pointer obtained
-      // from Box::into_raw, hence, satisfies memory layout requirements.
-      let _ = unsafe { Box::from_raw(sym) };
-      if let Some(fast_call_ptr) = fast_call_alloc {
-        // fast-call compiled trampoline is unmapped when the MMAP handle is dropped
-        // SAFETY: This is never called twice. pointer obtained
-        // from Box::into_raw, hence, satisfies memory layout requirements.
-        let _ = unsafe { Box::from_raw(fast_call_ptr) };
-      }
-    }),
-  );
-
-  weak.to_local(scope).unwrap()
-}
 
 #[inline]
 fn ffi_parse_bool_arg(
@@ -2069,462 +1650,4 @@ fn op_ffi_call_nonblocking<'scope>(
     // SAFETY: Same return type declared to libffi; trust user to have it right beyond that.
     Ok(unsafe { result.to_value(result_type) })
   })
-}
-
-#[op(fast)]
-fn op_ffi_ptr_of<FP>(
-  state: &mut deno_core::OpState,
-  buf: *const u8,
-  out: &mut [u32],
-) -> Result<(), AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointer#of");
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let outptr = out.as_ptr() as *mut usize;
-  let length = out.len();
-  assert!(
-    length >= (std::mem::size_of::<usize>() / std::mem::size_of::<u32>())
-  );
-  assert_eq!(outptr as usize % std::mem::size_of::<usize>(), 0);
-
-  // SAFETY: Out buffer was asserted to be at least large enough to hold a usize, and properly aligned.
-  let out = unsafe { &mut *outptr };
-  *out = buf as usize;
-
-  Ok(())
-}
-
-unsafe extern "C" fn noop_deleter_callback(
-  _data: *mut c_void,
-  _byte_length: usize,
-  _deleter_data: *mut c_void,
-) {
-}
-
-#[op(v8)]
-fn op_ffi_get_buf<FP, 'scope>(
-  scope: &mut v8::HandleScope<'scope>,
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-  len: usize,
-) -> Result<serde_v8::Value<'scope>, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#arrayBuffer");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *mut c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid FFI pointer value, got nullptr"));
-  }
-
-  // SAFETY: Offset is user defined.
-  let ptr = unsafe { ptr.add(offset) };
-
-  // SAFETY: Trust the user to have provided a real pointer, and a valid matching size to it. Since this is a foreign pointer, we should not do any deletion.
-  let backing_store = unsafe {
-    v8::ArrayBuffer::new_backing_store_from_ptr(
-      ptr,
-      len,
-      noop_deleter_callback,
-      std::ptr::null_mut(),
-    )
-  }
-  .make_shared();
-  let array_buffer: v8::Local<v8::Value> =
-    v8::ArrayBuffer::with_backing_store(scope, &backing_store).into();
-  Ok(array_buffer.into())
-}
-
-#[op(fast)]
-fn op_ffi_buf_copy_into<FP>(
-  state: &mut deno_core::OpState,
-  src: usize,
-  offset: usize,
-  dst: &mut [u8],
-  len: usize,
-) -> Result<(), AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#copyInto");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  if dst.len() < len {
-    Err(range_error(
-      "Destination length is smaller than source length",
-    ))
-  } else {
-    let src = src as *const c_void;
-
-    // SAFETY: Offset is user defined.
-    let src = unsafe { src.add(offset) as *const u8 };
-
-    // SAFETY: src is user defined.
-    // dest is properly aligned and is valid for writes of len * size_of::<T>() bytes.
-    unsafe { ptr::copy::<u8>(src, dst.as_mut_ptr(), len) };
-    Ok(())
-  }
-}
-
-#[op(v8)]
-fn op_ffi_cstr_read<FP, 'scope>(
-  scope: &mut v8::HandleScope<'scope>,
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<serde_v8::Value<'scope>, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getCString");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid CString pointer, pointer is null"));
-  }
-
-  // SAFETY: Offset is user defined.
-  let ptr = unsafe { ptr.add(offset) };
-
-  // SAFETY: Pointer is user provided.
-  let cstr = unsafe { CStr::from_ptr(ptr as *const c_char) }
-    .to_str()
-    .map_err(|_| type_error("Invalid CString pointer, not valid UTF-8"))?;
-  let value: v8::Local<v8::Value> = v8::String::new(scope, cstr)
-    .ok_or_else(|| {
-      type_error("Invalid CString pointer, string exceeds max length")
-    })?
-    .into();
-  Ok(value.into())
-}
-
-#[op(fast)]
-fn op_ffi_read_bool<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<bool, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getBool");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid bool pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe { ptr::read_unaligned::<bool>(ptr.add(offset) as *const bool) })
-}
-
-#[op(fast)]
-fn op_ffi_read_u8<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<u32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getUint8");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid u8 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe { ptr::read_unaligned::<u8>(ptr.add(offset) as *const u8) as u32 })
-}
-
-#[op(fast)]
-fn op_ffi_read_i8<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<i32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getInt8");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid i8 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe { ptr::read_unaligned::<i8>(ptr.add(offset) as *const i8) as i32 })
-}
-
-#[op(fast)]
-fn op_ffi_read_u16<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<u32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getUint16");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid u16 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe {
-    ptr::read_unaligned::<u16>(ptr.add(offset) as *const u16) as u32
-  })
-}
-
-#[op(fast)]
-fn op_ffi_read_i16<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<i32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getInt16");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid i16 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe {
-    ptr::read_unaligned::<i16>(ptr.add(offset) as *const i16) as i32
-  })
-}
-
-#[op(fast)]
-fn op_ffi_read_u32<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<u32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getUint32");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid u32 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe {
-    ptr::read_unaligned::<u32>(ptr.add(offset) as *const u32) as u32
-  })
-}
-
-#[op(fast)]
-fn op_ffi_read_i32<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<i32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getInt32");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid i32 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe {
-    ptr::read_unaligned::<i32>(ptr.add(offset) as *const i32) as i32
-  })
-}
-
-#[op]
-fn op_ffi_read_u64<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-  out: &mut [u32],
-) -> Result<(), AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getBigUint64");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let outptr = out.as_mut_ptr() as *mut u64;
-
-  assert!(
-    out.len() >= (std::mem::size_of::<u64>() / std::mem::size_of::<u32>())
-  );
-  assert_eq!((outptr as usize % std::mem::size_of::<u64>()), 0);
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid u64 pointer, pointer is null"));
-  }
-
-  let value =
-  // SAFETY: ptr and offset are user provided.
-    unsafe { ptr::read_unaligned::<u64>(ptr.add(offset) as *const u64) };
-
-  // SAFETY: Length and alignment of out slice were asserted to be correct.
-  unsafe { *outptr = value };
-  Ok(())
-}
-
-#[op(fast)]
-fn op_ffi_read_i64<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-  out: &mut [u32],
-) -> Result<(), AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getBigUint64");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let outptr = out.as_mut_ptr() as *mut i64;
-
-  assert!(
-    out.len() >= (std::mem::size_of::<i64>() / std::mem::size_of::<u32>())
-  );
-  assert_eq!((outptr as usize % std::mem::size_of::<i64>()), 0);
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid i64 pointer, pointer is null"));
-  }
-
-  let value =
-  // SAFETY: ptr and offset are user provided.
-    unsafe { ptr::read_unaligned::<i64>(ptr.add(offset) as *const i64) };
-  // SAFETY: Length and alignment of out slice were asserted to be correct.
-  unsafe { *outptr = value };
-  Ok(())
-}
-
-#[op(fast)]
-fn op_ffi_read_f32<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<f32, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getFloat32");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid f32 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe { ptr::read_unaligned::<f32>(ptr.add(offset) as *const f32) })
-}
-
-#[op(fast)]
-fn op_ffi_read_f64<FP>(
-  state: &mut deno_core::OpState,
-  ptr: usize,
-  offset: usize,
-) -> Result<f64, AnyError>
-where
-  FP: FfiPermissions + 'static,
-{
-  check_unstable(state, "Deno.UnsafePointerView#getFloat64");
-
-  let permissions = state.borrow_mut::<FP>();
-  permissions.check(None)?;
-
-  let ptr = ptr as *const c_void;
-
-  if ptr.is_null() {
-    return Err(type_error("Invalid f64 pointer, pointer is null"));
-  }
-
-  // SAFETY: ptr and offset are user provided.
-  Ok(unsafe { ptr::read_unaligned::<f64>(ptr.add(offset) as *const f64) })
-}
-
-#[cfg(test)]
-mod tests {
-  #[cfg(target_os = "windows")]
-  #[test]
-  fn test_format_error() {
-    use super::format_error;
-
-    // BAD_EXE_FORMAT
-    let err = dlopen::Error::OpeningLibraryError(
-      std::io::Error::from_raw_os_error(0x000000C1),
-    );
-    assert_eq!(
-      format_error(err, "foo.dll".to_string()),
-      "foo.dll is not a valid Win32 application.\r\n".to_string(),
-    );
-  }
 }
